+++ unstructured/ingest/connector/git.py
@@ -29,6 +29,8 @@ class SimpleGitConfig(BaseConnectorConfig):
     download_only: bool = False
     metadata_include: Optional[str] = None
     metadata_exclude: Optional[str] = None
+    partition_by_api: bool = False
+    partition_endpoint: str = "https://api.unstructured.io/general/v0/general"
     fields_include: str = "element_id,text,type,metadata"
     flatten_metadata: bool = False