+++ unstructured/ingest/main.py
@@ -130,6 +130,19 @@ class MainProcess:
     "Example: --metadata-exclude filename,page_number ",
 )
 @click.option(
+    "--partition-by-api",
+    is_flag=True,
+    default=False,
+    help="Use a remote API to partition the files."
+    " Otherwise, use the function from partition.auto",
+)
+@click.option(
+    "--partition-endpoint",
+    default="https://api.unstructured.io/general/v0/general",
+    help="If partitioning via api, use the following host. "
+    "Default: https://api.unstructured.io/general/v0/general",
+)
+@click.option(
     "--local-input-path",
     default=None,
     help="Path to the location in the local file system that will be processed.",
@@ -402,6 +415,8 @@ def main(
     fields_include,
     flatten_metadata,
     max_docs,
+    partition_by_api,
+    partition_endpoint,
     local_input_path,
     local_recursive,
     local_file_glob,
@@ -423,6 +438,13 @@ def main(
             "mutually exclusive with each other.",
         )
         sys.exit(1)
+    if (
+        not partition_by_api
+        and partition_endpoint != "https://api.unstructured.io/general/v0/general"
+    ):
+        logger.warning(
+            "Ignoring --partition-endpoint because --partition-by-api was not set",
+        )
     if (not preserve_downloads and not download_only) and download_dir:
         logger.warning(
             "Not preserving downloaded files but --download_dir is specified",
@@ -501,6 +523,8 @@ def main(
                     preserve_downloads=preserve_downloads,
                     metadata_include=metadata_include,
                     metadata_exclude=metadata_exclude,
+                    partition_by_api=partition_by_api,
+                    partition_endpoint=partition_endpoint,
                     fields_include=fields_include,
                     flatten_metadata=flatten_metadata,
                     download_only=download_only,
@@ -531,6 +555,8 @@ def main(
                     preserve_downloads=preserve_downloads,
                     metadata_include=metadata_include,
                     metadata_exclude=metadata_exclude,
+                    partition_by_api=partition_by_api,
+                    partition_endpoint=partition_endpoint,
                     fields_include=fields_include,
                     flatten_metadata=flatten_metadata,
                     download_only=download_only,
@@ -558,6 +584,8 @@ def main(
                     preserve_downloads=preserve_downloads,
                     metadata_include=metadata_include,
                     metadata_exclude=metadata_exclude,
+                    partition_by_api=partition_by_api,
+                    partition_endpoint=partition_endpoint,
                     fields_include=fields_include,
                     flatten_metadata=flatten_metadata,
                     download_only=download_only,
@@ -582,6 +610,8 @@ def main(
                 re_download=re_download,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
                 download_only=download_only,
@@ -606,6 +636,8 @@ def main(
                 re_download=re_download,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
                 download_only=download_only,
@@ -632,6 +664,8 @@ def main(
                 re_download=re_download,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
                 download_only=download_only,
@@ -669,6 +703,8 @@ def main(
                 re_download=re_download,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
                 download_only=download_only,
@@ -693,6 +729,8 @@ def main(
                 re_download=re_download,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
                 download_only=download_only,
@@ -717,6 +755,8 @@ def main(
                 re_download=re_download,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
                 download_only=download_only,
@@ -737,6 +777,8 @@ def main(
                 output_dir=structured_output_dir,
                 metadata_include=metadata_include,
                 metadata_exclude=metadata_exclude,
+                partition_by_api=partition_by_api,
+                partition_endpoint=partition_endpoint,
                 fields_include=fields_include,
                 flatten_metadata=flatten_metadata,
             ),